TssSession - Class updates and additions

+++ build.ps1
@@ -1,6 +1,7 @@
 ﻿param(
     $gallerykey
 )
+$publish = $true
 $beta = $true
 $moduleName = 'Thycotic.SecretServer'
 $staging = "C:\temp\staging\"
@@ -12,11 +13,10 @@ $imported = Import-Module .\src\Thycotic.SecretServer.psd1 -Force -PassThru
 
 try {
     Import-Module Pester
-    Invoke-Pester -Path "$PSScriptRoot\tests" -Output Minimal
+    Invoke-Pester -Path "$PSScriptRoot\tests" -Output Minimal -ErrorAction Stop
 } catch {
-    if ($_.FullQualifiedErrorId -contains 'PesterAssertionFailed') {
-        throw "Pester test failed: $_"
-    }
+    throw "Pester test failed: $_"
+    $publish = $false
 }
 
 $foundModule = Find-Module -Name $moduleName -AllowPrerelease:$beta
@@ -30,18 +30,19 @@ if ($foundModule.Version -ge $imported.Version) {
     Write-Host "Module Files:"
     Get-ChildItem $moduleTempPath -Recurse | Select-Object Directory, Name
 
-    Write-Host "Publishing $moduleName [$($imported.Version)] to PowerShell Gallery"
-    try {
-        Publish-Module -Path $moduleTempPath -NuGetApiKey $gallerykey
-        Write-Host "successfully published - mock"
-    } catch {
-        throw "Publish to PowerShell Gallery failed: $($_.Exception.Message)"
-    }
-
-    if ($?) {
-        Write-Host "Published to PowerShell Gallery"
-        Remove-Item -Recurse -Force $staging
-    } else {
-        throw "PowerShell Gallery Publish Failed"
+    if ($publish) {
+        try {
+            Write-Host "Publishing $moduleName [$($imported.Version)] to PowerShell Gallery"
+            Publish-Module -Path $moduleTempPath -NuGetApiKey $gallerykey
+            Write-Host "successfully published - mock"
+        } catch {
+            throw "Publish to PowerShell Gallery failed: $($_.Exception.Message)"
+        }
+        if ($?) {
+            Write-Host "Published to PowerShell Gallery"
+            Remove-Item -Recurse -Force $staging
+        } else {
+            throw "PowerShell Gallery Publish Failed"
+        }
     }
 }
